+++ docs/website_files/styles.css
@@ -18,7 +18,7 @@ GRID CONTAINER FOR NOTEBOOK LINK CARDS
     box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
     transition: transform 0.3s ease, box-shadow 0.3s ease;
     text-align: center;
-    border: 4px solid rgb(97, 179, 230);
+    border: 4px solid rgb(146, 191, 218);
 }
 
 /*
@@ -48,7 +48,7 @@ CSV DOWNLOAD BUTTONS
 
 .download-button {
   display: inline-block;
-  background-color: rgb(97, 179, 230);
+  background-color: rgb(146, 191, 218);
   color: white;
   padding: 10px 18px;
   font-size: 16px;
@@ -79,7 +79,7 @@ DATASET DESCRIPTION GRID
   background-color: #FFFFFF;
   border-radius: 10px;
   box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
-  border: 4px solid rgb(97, 179, 230);
+  border: 4px solid rgb(146, 191, 218);
 }
 
 .description-grid p {
@@ -114,13 +114,13 @@ Variable Table
   background-color: #FFFFFF;
   border-radius: 10px; 
   box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1); 
-  border: 4px solid rgb(97, 179, 230); 
+  border: 4px solid rgb(146, 191, 218); 
   font-family: sans-serif; 
   margin: 0; 
 }
 
 .styled-table th {
-  background-color: rgb(97, 179, 230); 
+  background-color: rgb(146, 191, 218); 
   color: white; 
   text-align: center; 
   font-weight: bold; 
@@ -141,7 +141,7 @@ Tag Filters
 .tag-filter-container {
   margin-bottom: 1rem;
   padding: 0.5rem;
-  border: 5px solid rgb(97, 179, 230);
+  border: 5px solid rgb(146, 191, 218);
   border-radius: 10px;
   background-color: #f9f9f9;
   display: flex;
@@ -166,7 +166,7 @@ Clear Filters button
 #clear-filters {
   padding: 0.6rem 1.2rem;
   margin: 1rem auto;
-  border: 3px solid rgb(97, 179, 230);
+  border: 3px solid rgb(146, 191, 218);
   background-color: #e8f5ef;
   color: #2d6a4f;
   font-weight: bold;
@@ -178,7 +178,7 @@ Clear Filters button
 }
 
 #clear-filters:hover {
-  background-color: rgb(190, 225, 247);
+  background-color: rgb(146, 191, 218);
   border-color: rgb(40, 148, 215);
 }
 
@@ -188,7 +188,7 @@ Banner and Banner Button
 
 .hero-banner {
   background: linear-gradient(
-    rgb(97, 179, 230),
+    rgb(146, 191, 218),
     #ffffff
   );
   background-size: cover;
@@ -211,7 +211,7 @@ Banner and Banner Button
 
 .hero-banner .btn {
   margin-top: 1rem;
-  background-color: rgb(97, 179, 230);
+  background-color: rgb(146, 191, 218);
   color: black;
   padding: 0.6rem 1.2rem;
   font-weight: bold;
@@ -225,9 +225,9 @@ Dataset Toggle
 #dataset-toggle {
   margin-bottom: 1rem;
   padding: 0.5rem;
-  border: 5px solid rgb(97, 179, 230);
+  border: 5px solid rgb(146, 191, 218);
   border-radius: 10px;
-  background-color: rgb(97, 179, 230);
+  background-color: rgb(146, 191, 218);
   display: flex;
   flex-direction: column;
   align-items: center;
